add isAuthenticated middleware to prevent unauthorized access to api routes
