+++ frontend/src/pages/ProjectViewPage/ProjectViewPage.tsx
@@ -27,7 +27,7 @@ const ProjectViewPage = () => {
         const response = await api.get(`/projects/${id}`);
         setProject(response.data);
       } catch (error) {
-        navigate("/projects/404");
+        navigate("/projects/notfound");
       }
     };
